+++ tests/models/test_ngsi_v2_iot.py
@@ -5,8 +5,8 @@ Test module for context broker models
 import unittest
 from typing import List
 
-from filip.models.ngsi_v2.iot import BaseAttribute, DeviceCommand, ServiceGroup, \
-    Device, TransportProtocol
+from filip.models.ngsi_v2.iot import DeviceCommand, ServiceGroup, \
+    Device, TransportProtocol, IoTABaseAttribute
 
 
 class TestContextv2IoTModels(unittest.TestCase):
@@ -37,16 +37,16 @@ class TestContextv2IoTModels(unittest.TestCase):
 
         # Test if all needed fields, detect all invalid strings
         for string in invalid_strings:
-            self.assertRaises(ValidationError, BaseAttribute,
+            self.assertRaises(ValidationError, IoTABaseAttribute,
                               name=string, type="name",
                               entity_name="name", entity_type="name")
-            self.assertRaises(ValidationError, BaseAttribute,
+            self.assertRaises(ValidationError, IoTABaseAttribute,
                               name="name", type=string,
                               entity_name="name", entity_type="name")
-            self.assertRaises(ValidationError, BaseAttribute,
+            self.assertRaises(ValidationError, IoTABaseAttribute,
                               name="name", type="name",
                               entity_name=string, entity_type="name")
-            self.assertRaises(ValidationError, BaseAttribute,
+            self.assertRaises(ValidationError, IoTABaseAttribute,
                               name="name", type="name",
                               entity_name="name", entity_type=string)
 
@@ -69,8 +69,8 @@ class TestContextv2IoTModels(unittest.TestCase):
 
         # Test if all needed fields, do not trow wrong errors
         for string in valid_strings:
-            BaseAttribute(name=string, type=string,
-                          entity_name=string, entity_type=string)
+            IoTABaseAttribute(name=string, type=string,
+                              entity_name=string, entity_type=string)
             DeviceCommand(name=string, type="name")
             ServiceGroup(entity_type=string, resource="", apikey="")
             Device(device_id="", entity_name=string, entity_type=string,
@@ -78,19 +78,19 @@ class TestContextv2IoTModels(unittest.TestCase):
 
         # Test for the special-string protected field if all strings are blocked
         for string in special_strings:
-            self.assertRaises(ValidationError, BaseAttribute,
-                              name=string, type="name",
-                              entity_name="name", entity_type="name")
-            self.assertRaises(ValidationError, BaseAttribute,
-                              name="name", type=string,
-                              entity_name="name", entity_type="name")
-            self.assertRaises(ValidationError,
-                              DeviceCommand, name=string, type="name")
+            with self.assertRaises(ValidationError):
+                IoTABaseAttribute(name=string, type="name", entity_name="name",
+                                  entity_type="name")
+            with self.assertRaises(ValidationError):
+                IoTABaseAttribute(name="name", type=string, entity_name="name",
+                                  entity_type="name")
+            with self.assertRaises(ValidationError):
+                DeviceCommand(name=string, type="name")
 
         # Test for the normal protected field if all strings are allowed
         for string in special_strings:
-            BaseAttribute(name="name", type="name",
-                          entity_name=string, entity_type=string)
+            IoTABaseAttribute(name="name", type="name",
+                              entity_name=string, entity_type=string)
             ServiceGroup(entity_type=string, resource="", apikey="")
             Device(device_id="", entity_name=string, entity_type=string,
                    transport=TransportProtocol.HTTP)