+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - move back to more simple docs design ([#32](https://github.com/RWTH-EBC/FiLiP/issues/32))
 - added MQTT notifications ([#24](https://github.com/RWTH-EBC/FiLiP/issues/24))
 - introduced [CHANGELOG.md](https://github.com/RWTH-EBC/FiLiP/blob/development/CHANGELOG.md) with versions
+- remodeled ngsi-v2 models ([#58,#59,#60](https://github.com/RWTH-EBC/FiLiP/issues/60))
 - improved ContextEntity and Device deleting methods ([#27](https://github.com/RWTH-EBC/FiLiP/issues/28))
 
 #### v0.1.7